feat(issues): Add organization tag values cmoponent
